refactor:테스트 함수 분리 및 명세 변경
