+++ src/domain/class/__test__/testUtil/mockRandomNumberInRange.js
@@ -1,0 +1,11 @@
+import getRandomNumberInRange from '../../../../utils/getRandomNumberInRange';
+
+const mockRandomNumberInRange = numbers => {
+  jest.mock('../../../../utils/getRandomNumberInRange');
+  console.log(getRandomNumberInRange);
+  numbers.forEach(n => {
+    getRandomNumberInRange.mockReturnValueOnce(n);
+  });
+};
+
+export default mockRandomNumberInRange;
